Sustainability portion of website beginning
This is for the website and also for you guys to see. I'll be adding more to sustainability and then to resiliency because these are global priorities that relate to all spectrums and professional fields. Just to get you to understand the concept of sustainability. Once you get a good grasp of the meaning, then try to implement it to a project, method, or problem in any particular field and create a sustainable solution if it is not already sustainable. I am happy to provide you with a scenario for practice. Hope this helps!
Enjoy, Tyler